remove frontend role selection for edit member
