--- atst/forms/edit_member.py
@@ -1,18 +1,0 @@
-from wtforms.validators import Required
-
-from .forms import BaseForm
-from atst.forms.fields import SelectField
-from atst.utils.localization import translate
-
-from .data import PORTFOLIO_ROLES
-
-
-class EditMemberForm(BaseForm):
-    # This form also accepts a field for each environment in each application
-    #  that the user is a member of
-
-    portfolio_role = SelectField(
-        translate("forms.edit_member.portfolio_role_label"),
-        choices=PORTFOLIO_ROLES,
-        validators=[Required()],
-    )
