+++ tests/domain/test_portfolios.py
@@ -114,10 +114,10 @@ def test_update_portfolio_role_role(portfolio, portfolio_owner):
     }
     PortfolioRoleFactory._meta.sqlalchemy_session_persistence = "flush"
     member = PortfolioRoleFactory.create(portfolio=portfolio)
-    role_name = "admin"
+    permission_sets = ["edit_portfolio_funding"]
 
     updated_member = Portfolios.update_member(
-        portfolio_owner, portfolio, member, role_name
+        portfolio_owner, portfolio, member, permission_sets=permission_sets
     )
     assert updated_member.portfolio == portfolio
